+++ src/yosys_run/run4.sh
@@ -2,4 +2,5 @@
 # whoami
 # $1 -h
 # yosys -p 'read_verilog top.v; read_verilog -D ICE40_HX -lib -specify +/ice40/cells_sim.v; hierarchy -check; proc; stat'
-$1 -p 'read_verilog top.v; hierarchy -simcheck -auto-top; tee -o out.json stat -json'
+# $1 -p 'read_verilog top.v; hierarchy -simcheck -auto-top; tee -o out.json stat -json'
+$1 -m slang -p 'read_slang top.v; hierarchy -simcheck -auto-top; tee -o out.json stat -json'